requirements.txt
discord.py
git+https://github.com/Rapptz/asqlite